Pulled function to its own page and updated it to be compatible with new solutions
